make a projectile interface
