+++ Source/BlockGame/Private/ProjectileInterface.cpp
@@ -1,0 +1,10 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+
+#include "ProjectileInterface.h"
+
+// Add default functionality here for any IPorjectileInterface functions that are not pure virtual.
+
+void IProjectileInterface::OnHit(UPrimitiveComponent* HitComp, AActor* OtherActor, UPrimitiveComponent* OtherComp, FVector NormalImpulse, const FHitResult& Hit)
+{
+}
