+++ Source/BlockGame/Public/Projectile.h
@@ -1,0 +1,40 @@
+﻿#pragma once
+
+#include "CoreMinimal.h"
+#include "ProjectileInterface.h"
+#include "GameFramework/Actor.h"
+#include "Projectile.generated.h"
+
+class USphereComponent;
+class UProjectileMovementComponent;
+
+UCLASS(Abstract, Blueprintable)
+class AProjectile : public AActor, public IProjectileInterface
+{
+	GENERATED_BODY()
+
+protected:
+	/** Sphere collision component */
+	UPROPERTY(VisibleDefaultsOnly, Category=Projectile)
+	USphereComponent* CollisionComp;
+
+	/** Projectile movement component */
+	UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category = Movement, meta = (AllowPrivateAccess = "true"))
+	UProjectileMovementComponent* ProjectileMovement;
+	
+public:
+	AProjectile() = default;
+
+	/** called when projectile hits something */
+	UFUNCTION()
+	virtual void OnHit(UPrimitiveComponent* HitComp, AActor* OtherActor, UPrimitiveComponent* OtherComp, FVector NormalImpulse, const FHitResult& Hit) override
+	{
+		
+	};
+
+	/** Returns CollisionComp subobject **/
+	USphereComponent* GetCollisionComp() const { return CollisionComp; }
+	/** Returns ProjectileMovement subobject **/
+	UProjectileMovementComponent* GetProjectileMovement() const { return ProjectileMovement; }
+};
+
